requirements.txt
pygame>=2.0
pillow
mutagen
requests
python-docx
pytest
pytest-xvfb
yt-dlp
pandas 
wikipedia 
beautifulsoup4
allure-pytest
kivy